server/adminapp/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.5
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.14.1
